requirements.txt
dj-database-url==0.5.0
Django==2.1.4
django-heroku==0.3.1
gunicorn==19.9.0
Jinja2==2.10
MarkupSafe==1.1.0
numpy==1.15.4
Pillow==5.3.0
psycopg2==2.7.3.2
pytesseract==0.2.6
pytz==2018.7
whitenoise==4.1.2
